% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-fEGarch_distr_est.R
\name{inf_criteria,fEGarch_distr_est-method}
\alias{inf_criteria,fEGarch_distr_est-method}
\alias{llhood,fEGarch_distr_est-method}
\alias{pars,fEGarch_distr_est-method}
\alias{se,fEGarch_distr_est-method}
\alias{vcov_mat,fEGarch_distr_est-method}
\title{Methods for Accessing Distribution Estimation Elements}
\usage{
\S4method{inf_criteria}{fEGarch_distr_est}(x)

\S4method{llhood}{fEGarch_distr_est}(x)

\S4method{pars}{fEGarch_distr_est}(x)

\S4method{se}{fEGarch_distr_est}(x)

\S4method{vcov_mat}{fEGarch_distr_est}(x)
}
\arguments{
\item{x}{an object returned by either \code{\link{distr_est}} or
its various wrappers like \code{\link{norm_est}}.}
}
\value{
The element within the input object of the same name as the method
is returned.
}
\description{
Accessors to access the elements of the same name in
output objects returned by either \code{\link{distr_est}} or
its various wrappers like \code{\link{norm_est}}.
}
\details{
Convenience methods to access the elements of the same name
that can otherwise be accessed via the operator \code{@} within
objects that inherit from class \code{"fEGarch_distr_est"}, which covers
objects returned by either \code{\link{distr_est}} or
its various wrappers like \code{\link{norm_est}}.
}
\examples{
x <- rged_s(4000, shape = 1.5) * 2.1 + 3.3
est <- ged_est(x)
inf_criteria(est)
pars(est)

}
