% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ufRisk-functions.R
\name{trafflight_test}
\alias{trafflight_test}
\alias{uncond_cov_test}
\alias{indep_test}
\alias{cond_cov_test}
\alias{cov_tests}
\alias{backtest_suite}
\alias{WAD}
\title{Generics for backtests}
\usage{
trafflight_test(object, ...)

uncond_cov_test(object, ...)

indep_test(object, ...)

cond_cov_test(object, ...)

cov_tests(object, ...)

backtest_suite(object, ...)

WAD(object, ...)
}
\arguments{
\item{object}{the generics are currently without use.}

\item{...}{the generics are currently without use.}
}
\value{
The generics are currently without use, can therefore not be called
and thus don't produce results.
}
\description{
Generic functions to build backtesting methods from.
}
\details{
The generics are currently without use.
}
