% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-egarch-spec.R
\name{fEGarch_fit,egarch_type_spec-method}
\alias{fEGarch_fit,egarch_type_spec-method}
\title{Fitting Method for Type I EGARCH-Family Models}
\usage{
\S4method{fEGarch_fit}{egarch_type_spec}(
  spec = egarch_spec(),
  rt,
  drange = c(0, 1),
  meanspec = mean_spec(),
  Drange = c(0, 1),
  n_test = 0,
  start_pars = NULL,
  LB = NULL,
  UB = NULL,
  control = list(),
  parallel = TRUE,
  ncores = max(1, future::availableCores() - 1),
  trunc = "none",
  presample = 50,
  Prange = c(1, 5)
)
}
\arguments{
\item{spec}{the generic is currently without use.}

\item{rt}{the generic is currently without use.}

\item{drange}{the generic is currently without use.}

\item{meanspec}{the generic is currently without use.}

\item{Drange}{the generic is currently without use.}

\item{n_test}{the generic is currently without use.}

\item{start_pars}{the generic is currently without use.}

\item{LB}{the generic is currently without use.}

\item{UB}{the generic is currently without use.}

\item{control}{the generic is currently without use.}

\item{parallel}{the generic is currently without use.}

\item{ncores}{the generic is currently without use.}

\item{trunc}{the generic is currently without use.}

\item{presample}{the generic is currently without use.}

\item{Prange}{the generic is currently without use.}
}
\value{
Returns an object of class \code{"fEGarch_fit_egarch"}.
}
\description{
Fits an EGARCH-family model of Type I. The method
is not being exported.
}
