% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-functions.R
\name{ljung_box_test}
\alias{ljung_box_test}
\alias{sign_bias_test}
\alias{goodn_of_fit_test}
\alias{fit_test_suite}
\title{Goodness-of-Fit Test Generics}
\usage{
ljung_box_test(
  object,
  m_max = 20,
  weight_f = function(lag) {
     m <- max(lag)
     (m + 1 - lag)/m
 },
  adj_df = NULL,
  silent = FALSE,
  ...
)

sign_bias_test(object, silent = FALSE, ...)

goodn_of_fit_test(object, n_bins = c(20, 30, 40, 50), silent = FALSE, ...)

fit_test_suite(
  object,
  args_lbt = list(),
  args_sbt = list(),
  args_goft = list(),
  silent = FALSE,
  ...
)
}
\arguments{
\item{object}{currently without use.}

\item{m_max}{currently without use.}

\item{weight_f}{currently without use.}

\item{adj_df}{currently without use.}

\item{silent}{currently without use.}

\item{...}{currently without use.}

\item{n_bins}{currently without use.}

\item{args_lbt}{currently without use.}

\item{args_sbt}{currently without use.}

\item{args_goft}{currently without use.}
}
\value{
The generics are currently without use and do not
return anything.
}
\description{
Generics for different goodness-of-fit tests. Currently
without use.
}
