% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{fitted_object_generics}
\alias{fitted_object_generics}
\alias{sigt}
\alias{sigt,ANY-method}
\alias{cmeans}
\alias{cmeans,ANY-method}
\alias{etat}
\alias{etat,ANY-method}
\alias{llhood}
\alias{llhood,ANY-method}
\alias{inf_criteria}
\alias{inf_criteria,ANY-method}
\alias{pars}
\alias{pars,ANY-method}
\alias{se}
\alias{se,ANY-method}
\alias{vcov_mat}
\alias{vcov_mat,ANY-method}
\title{Generics for Accessing Model Estimation Output Elements}
\usage{
sigt(x)

cmeans(x)

etat(x)

llhood(x)

inf_criteria(x)

pars(x)

se(x)

vcov_mat(x)
}
\arguments{
\item{x}{an object returned by either \code{\link{fEGarch}},
\code{\link{fiaparch}} or \code{\link{figarch}}, etc.}
}
\value{
These generics do not return anything. Inspect the methods based on
them for practical purpose.
}
\description{
Accessors for the output returned by the main fitting functions of
the \code{fEGarch} package. The generics themselves are without use.
}
\details{
These generics are without direct use. Consider the
specific methods based on them.
}
