% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{predict_roll}
\alias{predict_roll}
\title{Generics for Forecasts}
\usage{
predict_roll(
  object,
  step_size = 1,
  trunc = NULL,
  refit_after = NULL,
  steady_window = FALSE,
  parallel = TRUE,
  ncores = max(1, future::availableCores() - 1),
  fitting_args = list(),
  ...
)
}
\arguments{
\item{object}{the generics are currently without use.}

\item{step_size}{the generics are currently without use.}

\item{trunc}{the generics are currently without use.}

\item{refit_after}{the generics are currently without use.}

\item{steady_window}{the generics are currently without use.}

\item{parallel}{the generics are currently without use.}

\item{ncores}{the generics are currently without use.}

\item{fitting_args}{the generics are currently without use.}

\item{...}{the generics are currently without use.}
}
\value{
The generics do not work themselves and therefore
do not return anything.
}
\description{
The generics are themselves without use.
}
