% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting-functions.R
\name{fEGarch_predict,fEGarch_fit_egarch-method}
\alias{fEGarch_predict,fEGarch_fit_egarch-method}
\alias{fEGarch_predict,fEGarch_fit_loggarch-method}
\alias{fEGarch_predict,fEGarch_fit_aparch-method}
\alias{fEGarch_predict,fEGarch_fit_fiaparch-method}
\alias{fEGarch_predict,fEGarch_fit_gjrgarch-method}
\alias{fEGarch_predict,fEGarch_fit_figjrgarch-method}
\alias{fEGarch_predict,fEGarch_fit_tgarch-method}
\alias{fEGarch_predict,fEGarch_fit_fitgarch-method}
\alias{fEGarch_predict,fEGarch_fit_garch-method}
\alias{fEGarch_predict,fEGarch_fit_figarch-method}
\alias{predict_internal,fEGarch_fit-method}
\title{Prediction Methods for Package's Models}
\usage{
\S4method{fEGarch_predict}{fEGarch_fit_egarch}(object, n.ahead = 10, trunc = NULL, ...)

\S4method{fEGarch_predict}{fEGarch_fit_loggarch}(object, n.ahead = 10, trunc = NULL, ...)

\S4method{fEGarch_predict}{fEGarch_fit_aparch}(object, n.ahead = 10, trunc = NULL, ...)

\S4method{fEGarch_predict}{fEGarch_fit_fiaparch}(object, n.ahead = 10, trunc = NULL, ...)

\S4method{fEGarch_predict}{fEGarch_fit_gjrgarch}(object, n.ahead = 10, trunc = NULL, ...)

\S4method{fEGarch_predict}{fEGarch_fit_figjrgarch}(object, n.ahead = 10, trunc = NULL, ...)

\S4method{fEGarch_predict}{fEGarch_fit_tgarch}(object, n.ahead = 10, trunc = NULL, ...)

\S4method{fEGarch_predict}{fEGarch_fit_fitgarch}(object, n.ahead = 10, trunc = NULL, ...)

\S4method{fEGarch_predict}{fEGarch_fit_garch}(object, n.ahead = 10, trunc = NULL, ...)

\S4method{fEGarch_predict}{fEGarch_fit_figarch}(object, n.ahead = 10, trunc = NULL, ...)

\S4method{predict_internal}{fEGarch_fit}(object, n.ahead = 10, trunc = NULL, ...)
}
\arguments{
\item{object}{these methods are not being exported.}

\item{n.ahead}{these methods are not being exported.}

\item{trunc}{these methods are not being exported.}

\item{...}{these methods are not being exported.}
}
\value{
They return lists with two numeric vectors as elements named
\code{sigt} and \code{cmeans}.
}
\description{
Produces forecasts of the conditional standard deviation
(and of the conditional mean) following package's models.
These methods are not being exported and are used internally
only.
}
