% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-locpol_spec.R
\name{locpol_spec_methods}
\alias{locpol_spec_methods}
\alias{poly_order,locpol_spec-method}
\alias{kernel_order,locpol_spec-method}
\alias{boundary_method,locpol_spec-method}
\alias{bwidth,locpol_spec-method}
\alias{poly_order<-,locpol_spec-method}
\alias{kernel_order<-,locpol_spec-method}
\alias{boundary_method<-,locpol_spec-method}
\alias{bwidth<-,locpol_spec-method}
\title{Accessors for Class \code{"locpol_spec"}}
\usage{
\S4method{poly_order}{locpol_spec}(x)

\S4method{kernel_order}{locpol_spec}(x)

\S4method{boundary_method}{locpol_spec}(x)

\S4method{bwidth}{locpol_spec}(x)

\S4method{poly_order}{locpol_spec}(x) <- value

\S4method{kernel_order}{locpol_spec}(x) <- value

\S4method{boundary_method}{locpol_spec}(x) <- value

\S4method{bwidth}{locpol_spec}(x) <- value
}
\arguments{
\item{x}{the input object or object to modify.}

\item{value}{the value to modify the object \code{x} with.}
}
\value{
These methods return an object of class \code{"locpol_spec"}.
}
\description{
Access and change elements in objects of class \code{"locpol_spec"}.
The method names represent the name of the element to access /
manipulate.
}
\details{
These methods are intended to be used for accessing or manipulating
individual elements of objects of class \code{"locpol_spec"}.
}
\examples{
test_obj <- locpol_spec()
poly_order(test_obj)
poly_order(test_obj) <- 1
poly_order(test_obj)

}
