% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ufRisk-functions.R
\name{loss_functions}
\alias{loss_functions}
\title{Generic for Loss Function Calculation}
\usage{
loss_functions(object, penalty = 1e-04, ...)
}
\arguments{
\item{object}{currently without use.}

\item{penalty}{currently without use.}

\item{...}{currently without use.}
}
\value{
The generic itself is currently without use and thus
does not return anything.
}
\description{
Currently without use. Use the methods derived from
this generic.
}
