% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mean_spec.R
\name{mean_spec_methods}
\alias{mean_spec_methods}
\alias{orders,mean_spec-method}
\alias{long_memo,mean_spec-method}
\alias{include_mean,mean_spec-method}
\alias{orders<-,mean_spec-method}
\alias{long_memo<-,mean_spec-method}
\alias{include_mean<-,mean_spec-method}
\title{Accessors for Class \code{"mean_spec"}}
\usage{
\S4method{orders}{mean_spec}(x)

\S4method{long_memo}{mean_spec}(x)

\S4method{include_mean}{mean_spec}(x)

\S4method{orders}{mean_spec}(x) <- value

\S4method{long_memo}{mean_spec}(x) <- value

\S4method{include_mean}{mean_spec}(x) <- value
}
\arguments{
\item{x}{the input object or object to modify.}

\item{value}{the value to modify the object \code{x} with.}
}
\value{
These methods return an object of class \code{"mean_spec"}.
}
\description{
Access and change elements in objects of class \code{"mean_spec"}.
The method names represent the name of the element to access /
manipulate.
}
\details{
These methods are intended to be used for accessing or manipulating
individual elements of objects of class \code{"mean_spec"}.
}
\examples{
test_obj <- mean_spec()
orders(test_obj)
orders(test_obj) <- c(1, 1)
orders(test_obj)

}
