% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{poly_order}
\alias{poly_order}
\alias{kernel_order}
\alias{boundary_method}
\alias{bwidth}
\alias{poly_order<-}
\alias{kernel_order<-}
\alias{boundary_method<-}
\alias{bwidth<-}
\title{Generics for Nonparametric Smoothing Setting Adjustments}
\usage{
poly_order(x)

kernel_order(x)

boundary_method(x)

bwidth(x)

poly_order(x) <- value

kernel_order(x) <- value

boundary_method(x) <- value

bwidth(x) <- value
}
\arguments{
\item{x}{the generics are currently without use.}

\item{value}{the generics are currently without use.}
}
\value{
The generics do not work on their own and thus do not return anything.
}
\description{
Generics to build methods on for adjusting certain settings
in context of nonparametric smoothing. The generics are currently
without use.
}
