% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fEGarch_fit-plot.R
\name{plot}
\alias{plot}
\title{S4 Plot Generic}
\usage{
plot(x, y, ...)
}
\arguments{
\item{x}{see base R \code{plot} documentation.}

\item{y}{see base R \code{plot} documentation.}

\item{...}{see base R \code{plot} documentation.}
}
\value{
Returns nothing to the console but creates a plot in the plot window.
}
\description{
Imported from base R.
}
