% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-egarch-spec.R, R/aparchfit.R,
%   R/gjrgarchfit.R, R/tgarchfit.R, R/garchfit.R, R/fiaparchfit.R,
%   R/figjrgarchfit.R, R/fitgarchfit.R, R/figarchfit.R, R/rugarch-wrappers.R,
%   R/class-fEGarch_distr_est.R
\name{show,fEGarch_fit_egarch-method}
\alias{show,fEGarch_fit_egarch-method}
\alias{show,fEGarch_fit_loggarch-method}
\alias{show,fEgarch_fit_loggarch-method}
\alias{show,fEGarch_fit_aparch-method}
\alias{show,fEgarch_fit_aparch-method}
\alias{show,fEGarch_fit_gjrgarch-method}
\alias{show,fEgarch_fit_gjrgarch-method}
\alias{show,fEGarch_fit_tgarch-method}
\alias{show,fEgarch_fit_tgarch-method}
\alias{show,fEGarch_fit_garch-method}
\alias{show,fEgarch_fit_garch-method}
\alias{show,fEGarch_fit_fiaparch-method}
\alias{show,fEgarch_fit_fiaparch-method}
\alias{show,fEGarch_fit_figjrgarch-method}
\alias{show,fEgarch_fit_figjrgarch-method}
\alias{show,fEGarch_fit_fitgarch-method}
\alias{show,fEgarch_fit_fitgarch-method}
\alias{show,fEGarch_fit_figarch-method}
\alias{show,fEgarch_fit_figarch-method}
\alias{show,fEGarch_fit_rugarch_wrapper-method}
\alias{show,fEgarch_fit_rugarch_wrapper-method}
\alias{show,fEGarch_distr_est-method}
\alias{show,fEgarch_distr_est-method}
\title{Show Method for Estimation Output}
\usage{
\S4method{show}{fEGarch_fit_egarch}(object)

\S4method{show}{fEGarch_fit_loggarch}(object)

\S4method{show}{fEGarch_fit_aparch}(object)

\S4method{show}{fEGarch_fit_gjrgarch}(object)

\S4method{show}{fEGarch_fit_tgarch}(object)

\S4method{show}{fEGarch_fit_garch}(object)

\S4method{show}{fEGarch_fit_fiaparch}(object)

\S4method{show}{fEGarch_fit_figjrgarch}(object)

\S4method{show}{fEGarch_fit_fitgarch}(object)

\S4method{show}{fEGarch_fit_figarch}(object)

\S4method{show}{fEGarch_fit_rugarch_wrapper}(object)

\S4method{show}{fEGarch_distr_est}(object)
}
\arguments{
\item{object}{an object returned by one of this package's fitting functions,
for example \code{\link{fEGarch}},
\code{\link{fiaparch}}, \code{\link{figarch}}, etc.}
}
\value{
Nothing is returned. Results are printed in the R-console.
}
\description{
Display estimation results from the EGARCH family in
a convenient way in the console.
}
