% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{SP500}
\alias{SP500}
\title{Daily Log-Returns of the S&P 500}
\format{
An object of class \code{zoo} of length 6267.
}
\source{
The data was obtained from Yahoo Finance
(accessed: 2024-12-01) and then transformed into a time series
\code{"zoo"} object using R.

\url{https://finance.yahoo.com/quote/^GSPC/}
}
\usage{
SP500
}
\description{
A \code{"zoo"} object that contains the daily log-returns of the
S&P 500 (Standard and Poor's 500) index from January 04, 2000,
until November 30, 2024.
}
\keyword{datasets}
