% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{spec_generics}
\alias{spec_generics}
\alias{orders}
\alias{orders,ANY-method}
\alias{powers}
\alias{powers,ANY-method}
\alias{long_memo}
\alias{long_memo,ANY-method}
\alias{modulus}
\alias{modulus,ANY-method}
\alias{cond_dist}
\alias{cond_dist,ANY-method}
\alias{include_mean}
\alias{include_mean,ANY-method}
\alias{orders<-}
\alias{powers<-}
\alias{long_memo<-}
\alias{modulus<-}
\alias{cond_dist<-}
\alias{include_mean<-}
\title{Generics for Model Specification Accessors}
\usage{
orders(x)

powers(x)

long_memo(x)

modulus(x)

cond_dist(x)

include_mean(x)

orders(x) <- value

powers(x) <- value

long_memo(x) <- value

modulus(x) <- value

cond_dist(x) <- value

include_mean(x) <- value
}
\arguments{
\item{x}{the input object or object to modify.}

\item{value}{the value to modify the object \code{x} with.}
}
\value{
These generics themselves do not return anything and are just the
foundation for more sophisticated methods.
}
\description{
A collection of generics for accessors of the model specification
objects.
}
\details{
These generics are intended to provide a basis to construct methods
for adjusting the the output of model specification objects like
\code{egarch_spec()}.
}
