% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-egarch-spec.R
\name{spec_methods}
\alias{spec_methods}
\alias{orders,base_garch_spec-method}
\alias{powers,egarch_type_spec-method}
\alias{long_memo,base_garch_spec-method}
\alias{modulus,egarch_type_spec-method}
\alias{cond_dist,base_garch_spec-method}
\alias{orders<-,base_garch_spec-method}
\alias{powers<-,egarch_type_spec-method}
\alias{long_memo<-,base_garch_spec-method}
\alias{modulus<-,egarch_type_spec-method}
\alias{cond_dist<-,base_garch_spec-method}
\title{Accessors for Classes \code{"base_garch_spec"} and \code{"egarch_spec"}}
\usage{
\S4method{orders}{base_garch_spec}(x)

\S4method{powers}{egarch_type_spec}(x)

\S4method{long_memo}{base_garch_spec}(x)

\S4method{modulus}{egarch_type_spec}(x)

\S4method{cond_dist}{base_garch_spec}(x)

\S4method{orders}{base_garch_spec}(x) <- value

\S4method{powers}{egarch_type_spec}(x) <- value

\S4method{long_memo}{base_garch_spec}(x) <- value

\S4method{modulus}{egarch_type_spec}(x) <- value

\S4method{cond_dist}{base_garch_spec}(x) <- value
}
\arguments{
\item{x}{the input object or object to modify.}

\item{value}{the value to modify the object \code{x} with.}
}
\value{
These methods either return an object of class \code{"egarch_type_spec"}
or \code{"loggarch_type_spec"} or the corresponding element of
object of such class objects.
}
\description{
Access and change elements in objects of class \code{"egarch_spec"}.
The method names represent the name of the element to access /
manipulate.
}
\details{
These methods are intended to be used for accessing or manipulating
individual elements of objects of class \code{"egarch_spec"}.
}
\examples{
test_obj <- egarch_spec()
orders(test_obj)
orders(test_obj) <- c(2, 1)
orders(test_obj)

}
