% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{UKinflation}
\alias{UKinflation}
\title{Monthly Inflation Rate of the UK}
\format{
An object of class \code{ts} of length 540.
}
\source{
The data was obtained from the databank of the Federal
Reserve Bank of St. Louis
(accessed: 2025-06-17) and then transformed into a time series
\code{"ts"} object using R.

\url{https://fred.stlouisfed.org/series/CPALTT01GBM657N}
}
\usage{
UKinflation
}
\description{
A \code{"ts"} object that contains the monthly inflation rate of
the UK from January 1956 to December 2000. It is the growth rate
of the consumer price index from the previous period.
}
\keyword{datasets}
