% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{predict_roll}
\alias{predict_roll}
\title{Generics for Forecasts}
\usage{
predict_roll(object, step_size = 1, trunc = NULL, ...)
}
\arguments{
\item{object}{the generics are currently without use.}

\item{step_size}{the generics are currently without use.}

\item{trunc}{the generics are currently without use.}

\item{...}{the generics are currently without use.}
}
\value{
The generics do not work themselves and therefore
do not return anything.
}
\description{
The generics are themselves without use.
}
